+++ RenderSystems/GL3Plus/src/windowing/win32/OgreWin32Window.cpp
@@ -273,8 +273,8 @@ namespace Ogre {
                 adjustWindow(width, height, &winWidth, &winHeight);
 
                 // clamp window dimensions to screen size
-                int outerw = (winWidth < screenw)? winWidth : screenw;
-                int outerh = (winHeight < screenh)? winHeight : screenh;
+                int outerw = ((int)winWidth < screenw) ? (int)winWidth : screenw;
+                int outerh = ((int)winHeight < screenh) ? (int)winHeight : screenh;
 
                 if (left == -1)
                     left = monitorInfoEx.rcWork.left + (screenw - outerw) / 2;
@@ -600,8 +600,8 @@ namespace Ogre {
                 LONG screenh = monitorInfo.rcWork.bottom - monitorInfo.rcWork.top;
 
 
-                int left = screenw > winWidth ? ((screenw - winWidth) / 2) : 0;
-                int top = screenh > winHeight ? ((screenh - winHeight) / 2) : 0;
+                int left = (screenw > (int)winWidth) ? ((screenw - (int)winWidth) / 2) : 0;
+                int top = (screenh > (int)winHeight) ? ((screenh - (int)winHeight) / 2) : 0;
 
                 SetWindowLong(mHWnd, GWL_STYLE, getWindowStyle(mIsFullScreen));
                 SetWindowPos(mHWnd, HWND_NOTOPMOST, left, top, winWidth, winHeight,
